feat: Rendering from the server
