+++ src/app/page.tsx
@@ -1,103 +1,13 @@
-"use client";
+import { db } from "~/server/db";
+import {
+  files as fileSchema,
+  folders as folderSchema,
+} from "~/server/db/schema";
+import DriveContents from "./drive-contents";
 
-import { useMemo, useState } from "react";
-import { mockFiles, mockFolders } from "../lib/mock-data";
-import { Folder, FileIcon, Upload, ChevronRight } from "lucide-react";
-import Link from "next/link";
-import { Button } from "~/components/ui/button";
-import { FileRow, FolderRow } from "./file-row";
-
-export default function GoogleDriveClone() {
-  const [currentFolder, setCurrentFolder] = useState<string>("root");
-
-  const getCurrentFolders = () => {
-    return mockFolders.filter((file) => file.parent === currentFolder);
-  };
-
-  const getCurrentFiles = () => {
-    return mockFiles.filter((file) => file.parent === currentFolder);
-  };
-
-  const handleFolderClick = (folderId: string) => {
-    setCurrentFolder(folderId);
-  };
-
-  const breadcrumbs = useMemo(() => {
-    const breadcrumbs = [];
-    let currentId = currentFolder;
-
-    while (currentId !== "root") {
-      const folder = mockFolders.find((folder) => folder.id === currentId);
-      if (folder) {
-        breadcrumbs.unshift(folder);
-        currentId = folder.parent ?? "root";
-      } else {
-        break;
-      }
-    }
-
-    return breadcrumbs;
-  }, [currentFolder]);
-
-  const handleUpload = () => {
-    alert("Upload functionality would be implemented here");
-  };
-
-  return (
-    <div className="min-h-screen bg-gray-900 p-8 text-gray-100">
-      <div className="mx-auto max-w-6xl">
-        <div className="mb-6 flex items-center justify-between">
-          <div className="flex items-center">
-            <Button
-              onClick={() => setCurrentFolder("root")}
-              variant="ghost"
-              className="mr-2 text-gray-300 hover:text-white"
-            >
-              My Drive
-            </Button>
-            {breadcrumbs.map((folder) => (
-              <div key={folder.id} className="flex items-center">
-                <ChevronRight className="mx-2 text-gray-500" size={16} />
-                <Button
-                  onClick={() => handleFolderClick(folder.id)}
-                  variant="ghost"
-                  className="text-gray-300 hover:text-white"
-                >
-                  {folder.name}
-                </Button>
-              </div>
-            ))}
-          </div>
-          <Button
-            onClick={handleUpload}
-            className="bg-blue-600 text-white hover:bg-blue-700"
-          >
-            <Upload className="mr-2" size={20} />
-            Upload
-          </Button>
-        </div>
-        <div className="rounded-lg bg-gray-800 shadow-xl">
-          <div className="border-b border-gray-700 px-6 py-4">
-            <div className="grid grid-cols-12 gap-4 text-sm font-medium text-gray-400">
-              <div className="col-span-6">Name</div>
-              <div className="col-span-3">Type</div>
-              <div className="col-span-3">Size</div>
-            </div>
-          </div>
-          <ul>
-            {getCurrentFolders().map((folder) => (
-              <FolderRow
-                folder={folder}
-                key={folder.id}
-                handleFolderClick={() => handleFolderClick(folder.id)}
-              />
-            ))}
-            {getCurrentFiles().map((file) => (
-              <FileRow file={file} key={file.id} />
-            ))}
-          </ul>
-        </div>
-      </div>
-    </div>
-  );
+// SERVER COMPONENT
+export default async function GoogleDriveClone() {
+  const files = await db.select().from(fileSchema);
+  const folders = await db.select().from(folderSchema);
+  return <DriveContents files={files} folders={folders} />;
 }
